refactor accounts table, useEditAccount hook, and useSelectedId hook
