+++ hooks/use-selected-id.tsx
@@ -1,21 +1,21 @@
 import { useState } from 'react';
 
 /**
- * This hook keeps track of the currently selected ID. It can be used to control the open and close state of a component that shows some data related to the ID, for example, a modal component that contains a form to edit something.
+ * This hook keeps track of the currently selected ID. It can be used to control the state of a component that shows some data based on the selected ID, for example, the open and close state of a modal component that contains a form to edit a post.
  */
 
 export function useSelectedId() {
 	const [selectedId, setSelectedId] = useState<undefined | number>(undefined);
 
-	const isOpen = selectedId !== undefined;
+	const isSelectedId = selectedId !== undefined;
 
-	function handleOpen(id: number) {
+	function setId(id: number) {
 		setSelectedId(id);
 	}
 
-	function handleClose() {
+	function unsetId() {
 		setSelectedId(undefined);
 	}
 
-	return { isOpen, selectedId, handleOpen, handleClose };
+	return { isSelectedId, selectedId, setId, unsetId };
 }
